+++ lambda.reputation-list.tf
@@ -95,6 +95,7 @@ resource "aws_lambda_function" "reputation-list" {
   memory_size      = 512
   timeout          = 300
   publish          = true
+  layers           = [data.aws_ssm_parameter.powertools_layer.value]
 
   dead_letter_config {
     target_arn = var.dead_letter_arn